# web/documentation/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Yii2-user-api",
        "version": "0.1"
    },
    "paths": {
        "/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "description": "Get list of products",
                "operationId": "a291206236f4750b3866c3ca3635ac7f",
                "responses": {
                    "200": {
                        "description": "Products list"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "operationId": "9744a79b9e7b09d3251df79bd7caa278",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "date_and_time"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "string"
                                    },
                                    "date_and_time": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Product",
            "description": "Products endpoints"
        }
    ]
}